+++ internal/domain/category/repository/mock.go
@@ -5,6 +5,7 @@ import (
 
 	"personal-finance/internal/model"
 
+	"github.com/google/uuid"
 	"github.com/stretchr/testify/mock"
 )
 
@@ -12,27 +13,27 @@ type Mock struct {
 	mock.Mock
 }
 
-func (m *Mock) Add(_ context.Context, category model.Category, userID string) (model.Category, error) {
-	args := m.Called(category, userID)
+func (m *Mock) Add(_ context.Context, category model.Category) (model.Category, error) {
+	args := m.Called(category)
 	return args.Get(0).(model.Category), args.Error(1)
 }
 
-func (m *Mock) FindAll(_ context.Context, userID string) ([]model.Category, error) {
-	args := m.Called(userID)
+func (m *Mock) FindAll(_ context.Context) ([]model.Category, error) {
+	args := m.Called()
 	return args.Get(0).([]model.Category), args.Error(1)
 }
 
-func (m *Mock) FindByID(_ context.Context, id int, userID string) (model.Category, error) {
-	args := m.Called(id, userID)
+func (m *Mock) FindByID(_ context.Context, id uuid.UUID) (model.Category, error) {
+	args := m.Called(id)
 	return args.Get(0).(model.Category), args.Error(1)
 }
 
-func (m *Mock) Update(_ context.Context, id int, category model.Category, userID string) (model.Category, error) {
-	args := m.Called(id, category, userID)
+func (m *Mock) Update(_ context.Context, id uuid.UUID, category model.Category) (model.Category, error) {
+	args := m.Called(id, category)
 	return args.Get(0).(model.Category), args.Error(1)
 }
 
-func (m *Mock) Delete(_ context.Context, _ int) error {
-	args := m.Called()
+func (m *Mock) Delete(_ context.Context, id uuid.UUID) error {
+	args := m.Called(id)
 	return args.Error(0)
 }